refact(scanner): count in int32, the type a token conveys
token.Position carries lines, columns, offsets and indentation as int32,
and the scanner counted them in int, so every position narrowed on its
way into a token. The counters are int32 now -- ten on Scanner, six on
Context, six on MultiLineState, and propertyRun.length -- and token.At
takes the fields as they stand.

    Scanner         720 -> 640
    Context         424 -> 376
    MultiLineState   88 ->  64
    propertyRun      16 ->  12

Both are long-lived: one Scanner and one Context per document, held for
as long as the source is read.

The narrowing did not vanish, it moved: len and utf8.RuneCount hand back
int, so 62 sites now convert at that edge instead of 11 converting at the
token. Each is bounded by the length of the source, which Init already
caps below math.MaxInt32, and .golangci.yml excludes G115 for the package
with that argument written down. What the exclusion costs is a narrowing
in code not yet written that nobody has argued is bounded.

Measured a wash: +0.23% per token geomean over the twenty-six shapes of
BenchmarkScannerNextToken, interleaved, n=10, with the two significant
shapes pointing opposite ways. Truncation is free and the extension an
index needs is the one the load already does.

Context.read stays an int: it indexes pending, which holds two tokens,
and never reaches a Position.

+++ internal/scanner/anchor.go
@@ -59,7 +59,7 @@ func (s *Scanner) validateAnchorName(ctx *Context, what string) error {
 	switch {
 	case end == start:
 		return ErrInvalidToken(what+" must be followed by a name", token.Invalid(ctx.origin(), s.pos()))
-	case end < len(ctx.src) && (ctx.src[end] == '[' || ctx.src[end] == '{'):
+	case end < int32(len(ctx.src)) && (ctx.src[end] == '[' || ctx.src[end] == '{'):
 		return ErrInvalidToken(what+" must be separated from the node that follows it", token.Invalid(ctx.origin(), s.pos()))
 	default:
 		return nil
@@ -71,10 +71,10 @@ func (s *Scanner) validateAnchorName(ctx *Context, what string) error {
 // ns-anchor-char is ns-char less the flow indicators, so a name runs up to whitespace, a line break, the end of the
 // input, or one of ',', '[', ']', '{' or '}'.
 // A ':' is none of those and belongs to the name, which is why "{&a: b}" anchors a node named "a:".
-func anchorNameEnd(src string, start int) int {
+func anchorNameEnd(src string, start int32) int32 {
 	// Every character that ends a name is ASCII, so this can walk bytes: no byte of a multi-byte character is one of them.
 	end := start
-	for end < len(src) && !endsAnchorName(rune(src[end])) {
+	for end < int32(len(src)) && !endsAnchorName(rune(src[end])) {
 		end++
 	}
