+++ internal/scanner/blanks.go
@@ -153,7 +153,7 @@ const (
 // stand equal where the line is genuinely opening.
 // Where they do not, characters have been read on this line by a path that never reached updateIndent -- a block
 // scalar's content, a quoted scalar spanning a break -- and the counts this advances in step are already apart.
-func (s *Scanner) indentRun(ctx *Context) int {
+func (s *Scanner) indentRun(ctx *Context) int32 {
 	if s.indentNum != s.column {
 		return 0
 	}
@@ -161,7 +161,7 @@ func (s *Scanner) indentRun(ctx *Context) int {
 	raw := ctx.raw
 	i := ctx.idx
 	if !s.deepIndent {
-		probe := min(i+indentProbe, len(raw))
+		probe := min(i+indentProbe, int32(len(raw)))
 		for ; i < probe; i++ {
 			if raw[i] != ' ' {
 				return i - ctx.idx
@@ -170,15 +170,15 @@ func (s *Scanner) indentRun(ctx *Context) int {
 	}
 
 	// Eight bytes at a time: the run outran the probe, or the document has already shown that its lines are indented.
-	for i+8 <= len(raw) {
+	for i+8 <= int32(len(raw)) {
 		w := binary.LittleEndian.Uint64(raw[i:])
 		if m := swar.SpaceMask(w); m != 0 {
-			return i + swar.FirstByte(m) - ctx.idx
+			return i + int32(swar.FirstByte(m)) - ctx.idx
 		}
 		i += 8
 	}
 
-	for i < len(raw) && raw[i] == ' ' {
+	for i < int32(len(raw)) && raw[i] == ' ' {
 		i++
 	}
